LayoutTests/media/media-webm-invalid-check-expected.txt


RUN(video.src = "content/invalid-size.webm")
EVENT(error)
Properly failed to load media OK
END OF TEST
